Move new method to file dedicated to layers
